refactor module structure update poetry
