+++ adapters/utils/id_type_processing.py
