+++ adapters/utils/transactions.py
